requirements.txt
flask[async]
openai
edge-tts
pydub
markdown2
coqui-tts